refactor: add changes according to review
