+++ my-app/src/entities/post/ui/PosstCard.module.css
@@ -8,11 +8,12 @@
   align-items: center;
   border: solid 1px rgb(57, 36, 36);
   margin-bottom: 1rem;
+  background-color: var(--card-bg);
 }
 .cardTitle {
   font-size: 2rem;
   font-weight: 600;
-  color: rgb(15, 15, 55);
+  color: var(--text-color);
 }
 .imgWrap {
   width: 100%;
@@ -27,7 +28,7 @@ img {
 .cardDescription {
   font-size: 1rem;
   font-weight: 400;
-  color: rgb(21, 21, 88);
+  color: var(--text-color);
 }
 .imgDefault {
   background-color: rgb(137, 48, 102);
