+++ my-app/src/features/ThemeSwitcher/ui/ThemeSwitcher.tsx
@@ -1,4 +1,4 @@
-import { useTheme } from "../../../shared/lib/theme/ThemeContext";
+import { useTheme } from "@/shared/lib/theme/use-theme";
 import { Button } from "../../../shared/ui/Button/Button";
