+++ my-app/src/main.tsx
@@ -3,7 +3,7 @@ import { StrictMode } from 'react'
 import { createRoot } from 'react-dom/client'
 import './index.css'
 import App from './app/App'
-import { ThemeProvider } from './shared/lib/theme/ThemeContext'
+import { ThemeProvider } from './shared/lib/theme/ThemeProvider'
 
 createRoot(document.getElementById('root')!).render(
   <StrictMode>
