+++ my-app/src/shared/lib/theme/use-theme.tsx
@@ -1,0 +1,10 @@
+import { useContext } from "react";
+import { ThemeContext } from "./ThemeContext";
+
+export const useTheme = () => {
+  const context = useContext(ThemeContext);
+  if (!context) {
+    throw new Error("oops");
+  }
+  return context;
+};
