+++ my-app/src/widgets/LayoutHeader/Header.module.css
@@ -1,9 +1,17 @@
 header {
   padding: 0.75rem;
-  background-color: rgb(189, 171, 206);
+  background-color: var(--header-bg);
+  color: var(--text-color);
 }
 .container {
   display: flex;
   justify-content: space-between;
+  align-items: center;
+}
+.rightWrap {
+  display: flex;
+  justify-content: center;
+  gap: 0.5rem;
+  width: 25%;
   align-items: center;
 }
