+++ my-app/src/widgets/LayoutHeader/Header.tsx
@@ -1,11 +1,15 @@
+import { ThemeSwitcher } from '@/features/ThemeSwitcher/ui/ThemeSwitcher';
 import styles from './Header.module.css'
 export const Header = () => {
-  const { container } = styles;
+  const { container, rightWrap } = styles;
   return (
     <header>
       <div className={container}>
         <p>My-app</p>
-        <p>How you doin'?</p>
+        <div className={rightWrap}>
+          <ThemeSwitcher/>
+          <p>How you doin'?</p>
+        </div>
       </div>
     </header>
   )
